# api/openapi/internal-v1.openapi.yaml
openapi: 3.1.0
info:
  title: AFSCP Internal API
  version: 1.0.0
  description: Product-agnostic internal storage control plane API.
  license:
    name: Proprietary internal
    identifier: LicenseRef-Proprietary
servers:
  - url: https://afscp.internal
security:
  - serviceAuth: []
components:
  securitySchemes:
    serviceAuth:
      type: http
      scheme: bearer
      description: Deployment service credential or mTLS-derived bearer facade.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema:
        type: string
    ActorType:
      name: X-AFSCP-Actor-Type
      in: header
      required: true
      schema:
        type: string
    ActorId:
      name: X-AFSCP-Actor-Id
      in: header
      required: true
      schema:
        type: string
    CallerService:
      name: X-AFSCP-Caller-Service
      in: header
      required: true
      schema:
        type: string
    NamespaceId:
      name: X-AFSCP-Namespace-Id
      in: header
      required: true
      description: Required namespace context for namespace-bound requests. When the request also carries a namespace path, query, or body field, values must match.
      schema:
        $ref: ../schemas/afscp-internal-v1.schema.json#/$defs/NamespaceId
  schemas:
    OperationEnvelope:
      $ref: ../schemas/afscp-internal-v1.schema.json#/$defs/OperationEnvelope
    ErrorEnvelope:
      $ref: ../schemas/afscp-internal-v1.schema.json#/$defs/ErrorEnvelope
...
    NamespaceVolumeBinding:
      $ref: ../schemas/afscp-internal-v1.schema.json#/$defs/NamespaceVolumeBinding
...
    RestoreRequest:
      $ref: ../schemas/afscp-internal-v1.schema.json#/$defs/RestoreRequest
...
  responses:
    Error:
      description: Standard error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
paths:
...
  /internal/v1/namespaces/{namespaceId}/volume-binding:
    put:
      operationId: putNamespaceVolumeBinding
      summary: Upsert namespace volume binding and policy.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CallerService"
        - $ref: "#/components/parameters/NamespaceId"
        - $ref: "#/components/parameters/ActorType"
        - $ref: "#/components/parameters/ActorId"
        - name: namespaceId
          in: path
          required: true
          schema:
            $ref: ../schemas/afscp-internal-v1.schema.json#/$defs/NamespaceId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NamespaceVolumeBinding"
      responses:
        "400":
          $ref: "#/components/responses/Error"
        "200":
          description: Binding upserted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationEnvelope"
        default:
          $ref: "#/components/responses/Error"
    get:
      operationId: getNamespaceVolumeBinding
      summary: Get namespace volume binding and policy.
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CallerService"
        - $ref: "#/components/parameters/NamespaceId"
        - name: namespaceId
          in: path
          required: true
          schema:
            $ref: ../schemas/afscp-internal-v1.schema.json#/$defs/NamespaceId
      responses:
        "400":
          $ref: "#/components/responses/Error"
        "200":
          description: Binding.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamespaceVolumeBinding"
        default:
          $ref: "#/components/responses/Error"
...
  /internal/v1/repos/{repoId}/restore:
    post:
      operationId: restore
      summary: Directly restore a repo to a save point.
      description: Creates a durable direct restore operation. This endpoint does not create a planning artifact, secondary run request, cleanup request, or safety save point.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CallerService"
        - $ref: "#/components/parameters/NamespaceId"
        - $ref: "#/components/parameters/ActorType"
        - $ref: "#/components/parameters/ActorId"
        - name: repoId
          in: path
          required: true
          schema:
            $ref: ../schemas/afscp-internal-v1.schema.json#/$defs/RepoId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreRequest"
      responses:
        "400":
          $ref: "#/components/responses/Error"
        "202":
          description: Direct restore operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationEnvelope"
        default:
          $ref: "#/components/responses/Error"
...